+++ example/lib/serialization.freezed.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// Hand-written to mimic freezed output. NOT produced by build_runner.
+// ignore_for_file: type=lint
+
+part of 'serialization.dart';
+
+mixin _$Point {
+  int get x;
+  int get y;
+  Map<String, dynamic> toJson();
+}
+
+class _Point extends Point {
+  const _Point(this.x, this.y) : super._();
+  @override
+  final int x;
+  @override
+  final int y;
+  @override
+  Map<String, dynamic> toJson() => _$PointToJson(this);
+}